+++ example/lib/repositories/people_repository.dart
@@ -1,121 +1,84 @@
 import 'package:clock/clock.dart';
 import 'package:cornerstone/cornerstone.dart';
 import 'package:dartz/dartz.dart';
-import 'package:equatable/equatable.dart';
 import 'package:example/data_sources/people_data_source.dart';
 import 'package:example/entities/person.dart';
-import 'package:example/repositories/hive_persistence_repository_mixin.dart';
 import 'package:hive/hive.dart';
 import 'package:json_annotation/json_annotation.dart';
-import 'package:meta/meta.dart';
 
 part 'people_repository.g.dart';
 
-abstract class PeopleRepository
-    with
-        LocallyPersistentRepository<PeopleSnapshot>,
-        HivePersistenceRepositoryMixin<PeopleSnapshot> {
-  Future<Either<Failure, PeopleSnapshot>> getPeople();
-}
-
-/// Remember to use `anyMap: true` if you use Hive because otherwise will throw:
-///
-/// ```text
-/// type '_InternalLinkedHashMap<dynamic, dynamic>' is not a subtype of type
-/// 'Map<String, dynamic>' in type cast
-/// ```
-@JsonSerializable(explicitToJson: true, checked: true, anyMap: true)
-class PeopleSnapshot extends Equatable {
-  @JsonKey(fromJson: _$dateTimeFromJson, toJson: _$dateTimeToJson)
-  final DateTime updatedAt;
-
+@JsonSerializable(explicitToJson: true, anyMap: true)
+class PeopleSnapshot extends CornerstoneSnapshot {
   @JsonKey(defaultValue: [])
   final List<Person> data;
 
-  @JsonKey(defaultValue: false)
-  final bool isSaved;
-
-  const PeopleSnapshot({
-    this.updatedAt,
+  PeopleSnapshot({
+    required DateTime timestamp,
     this.data = const [],
-    this.isSaved = false,
-  });
+    Clock clock = const Clock(),
+  }) : super(clock: clock, timestamp: timestamp);
+
+  @override
+  List<Object?> get props => [timestamp, data];
 
   factory PeopleSnapshot.fromJson(Map json) => _$PeopleSnapshotFromJson(json);
 
   Map<String, dynamic> toJson() => _$PeopleSnapshotToJson(this);
+}
 
-  @override
-  List<Object> get props => [updatedAt, isSaved, data];
+/// Equivalent to [PeopleRepository] but this time, the Hive functionality is
+/// built-in.
+abstract class PeopleRepository
+    with
+        LocallyPersistentRepository<PeopleSnapshot>,
+        CornerstonePersistentRepositoryMixin<PeopleSnapshot> {
+  Future<Either<Failure, PeopleSnapshot>> getPeople();
 }
 
 class PeopleRepositoryImpl extends PeopleRepository {
   final PeopleDataSource dataSource;
-  final HiveInterface hive;
   final Clock clock;
+  final HiveInterface hive;
+  final ConvertToFailure<Object?> convertToFailure;
+  final ConvertToSnapshot<PeopleSnapshot> convertToSnapshot;
 
-  PeopleSnapshot data = PeopleSnapshot();
+  @override
+  PeopleSnapshot snapshot;
 
-  Map<String, dynamic> get asJson => data.toJson();
+  Map<String, dynamic> get asJson => snapshot.toJson();
 
   PeopleRepositoryImpl({
-    @required this.dataSource,
-    @required this.hive,
+    required this.dataSource,
+    required this.hive,
+    required this.convertToSnapshot,
+    this.convertToFailure = const ConvertCornerstoneExceptionToFailure(),
     this.clock = const Clock(),
-  });
+  }) : snapshot = PeopleSnapshot(timestamp: clock.now(), clock: clock);
 
   @override
   Future<Either<Failure, PeopleSnapshot>> getPeople() async {
     try {
       final results = await dataSource.readMany();
 
-      data = PeopleSnapshot(
+      snapshot = PeopleSnapshot(
         data: results,
-        updatedAt: clock.now(),
-        isSaved: true,
+        timestamp: clock.now(),
+        clock: clock,
       );
 
-      final saveResult = await save();
+      await save();
 
-      if (saveResult.isLeft()) {
-        data = PeopleSnapshot(data: data.data, updatedAt: data.updatedAt);
-      }
-
-      return Right(data);
+      return Right(snapshot);
     } catch (e) {
-      return Left(Failure(
-        name: 'FAILED_TO_RETRIEVE_DATA',
-        details: e.toString(),
-      ));
-    }
-  }
-
-  @override
-  Future<Either<Failure, PeopleSnapshot>> load() async {
-    try {
-      final box = await hive.openBox(storageName);
-
-      /// For some reason, this is the only safe way I can get box data as
-      /// Map<String, dynamic> consistently. Using cast throws an error.
-      ///
-      /// See this [issue](https://github.com/hivedb/hive/issues/522) if you
-      /// want to get updated on this.
-      final result = box.toMap().map((k, e) => MapEntry(k.toString(), e));
-      data = PeopleSnapshot.fromJson(result);
-
-      return Right(data);
-    } on HiveError catch (e) {
-      return Left(Failure(name: e.message));
-    } catch (e) {
-      return Left(Failure(name: 'UNEXPECTED_ERROR', details: e.toString()));
+      return Left(convertToFailure(e));
     }
   }
 }
 
-/// Parse string date to DateTime object
-DateTime _$dateTimeFromJson(String date) =>
-    date != null && date.isNotEmpty ? DateTime.parse(date).toLocal() : null;
-
-/// Parse DateTime object to String
-String _$dateTimeToJson(DateTime date) =>
-    date != null ? date.toUtc().toIso8601String() : null;
+class ConvertToPeopleSnapshot implements ConvertToSnapshot<PeopleSnapshot> {
+  const ConvertToPeopleSnapshot();
+  @override
+  PeopleSnapshot call(Map<String, dynamic> data) =>
+      PeopleSnapshot.fromJson(data);
+}