+++ example/lib/use_cases/clear_people_storage.dart
@@ -3,15 +3,14 @@ import 'dart:async';
 import 'package:cornerstone/cornerstone.dart';
 import 'package:dartz/dartz.dart';
 import 'package:example/repositories/people_repository.dart';
-import 'package:meta/meta.dart';
 
 /// Clears local storage of people
-class ClearPeopleStorage extends UseCase<Failure, Unit, Null> {
+class ClearPeopleStorage extends UseCase<Failure, Unit, Unit> {
   final PeopleRepository repo;
 
-  ClearPeopleStorage({@required this.repo});
+  ClearPeopleStorage({required this.repo});
   @override
-  FutureOr<Either<Failure, Unit>> call({Null param}) async {
+  FutureOr<Either<Failure, Unit>> call({Unit param = unit}) async {
     return await repo.clear();
   }
 }