+++ lib/src/core/data_source.dart
@@ -1,5 +1,8 @@
 import 'dart:async';
 
+import 'package:cornerstone/cornerstone.dart';
+import 'package:dartz/dartz.dart';
+
 /// Implement this abstract in your data source if it can fetch some data.
 /// Note that 1 class can **implements** from multiple abstracts.
 ///
@@ -13,7 +16,7 @@ import 'dart:async';
 ///
 ///   final HttpClient client;
 ///
-///   const SessionDataSource({@required this.client});
+///   const SessionDataSource({required this.client});
 ///
 ///   @override
 ///   FutureOr<Session> readOne({GetParams params}) async {
@@ -27,7 +30,12 @@ import 'dart:async';
 ///
 /// Instead of throwing an exception, you should return Left instead.
 abstract class SingleGetterDataSource<Type, Param> {
-  FutureOr<Type> readOne({Param param});
+  FutureOr<Type> readOne({required Param param});
+}
+
+/// Similar to [SingleGetterDataSource], only returns Either form right away.
+abstract class SafeSingleGetterDataSource<Type, Param> {
+  FutureOr<Either<Failure, Type>> readOne({required Param param});
 }
 
 /// Implement this abstract in your data source if it can fetch some data.
@@ -43,7 +51,7 @@ abstract class SingleGetterDataSource<Type, Param> {
 ///
 ///   final HttpClient client;
 ///
-///   const PeopleDataSource({@required this.client});
+///   const PeopleDataSource({required this.client});
 ///
 ///   @override
 ///   FutureOr<List<Person>> readMany({GetParams params}) async {
@@ -55,7 +63,12 @@ abstract class SingleGetterDataSource<Type, Param> {
 /// If you can live with less type safety, you can simply use
 /// `Map<String,dynamic>` as Param.
 abstract class MultipleGetterDataSource<Type, Param> {
-  FutureOr<List<Type>> readMany({Param param});
+  FutureOr<List<Type>> readMany({required Param param});
+}
+
+/// Similar to [MultipleGetterDataSource], only returns Either form right away.
+abstract class SafeMultipleGetterDataSource<Type, Param> {
+  FutureOr<Either<Failure, List<Type>>> readMany({required Param param});
 }
 
 /// Implement this abstract in your data source if it can perform data creation.
@@ -69,7 +82,7 @@ abstract class MultipleGetterDataSource<Type, Param> {
 /// class AuthDataSource implements CreatorDataSource<Session, SignInParams> {
 ///   final HttpClient client;
 ///
-///   const AuthDataSource({@required this.client});
+///   const AuthDataSource({required this.client});
 ///
 ///   @override
 ///   FutureOr<Person> read({GetParams params}) async {
@@ -81,7 +94,12 @@ abstract class MultipleGetterDataSource<Type, Param> {
 /// If you can live with less type safety, you can simply use
 /// `Map<String,dynamic>` as Param.
 abstract class CreatorDataSource<Type, Param> {
-  FutureOr<Type> create({Param param});
+  FutureOr<Type> create({required Param param});
+}
+
+/// Similar to [CreatorDataSource], only returns Either form right away.
+abstract class SafeCreatorDataSource<Type, Param> {
+  FutureOr<Either<Failure, Type>> create({required Param param});
 }
 
 /// Implement this abstract in your data source if it can perform data update
@@ -91,10 +109,10 @@ abstract class CreatorDataSource<Type, Param> {
 /// class PeopleDataSource implements UpdaterDataSource<Person, Person> {
 ///   final HttpClient client;
 ///
-///   const PeopleDataSource({@required this.client});
+///   const PeopleDataSource({required this.client});
 ///
 ///   @override
-///   FutureOr<Person> read({Person params}) async {
+///   FutureOr<Person> read({required Person params}) async {
 ///     /// Perform client.put() or something using json form of provided
 ///     /// Person object.
 ///   }
@@ -104,20 +122,26 @@ abstract class CreatorDataSource<Type, Param> {
 /// If you can live with less type safety, you can simply use
 /// `Map<String,dynamic>` as Param.
 abstract class UpdaterDataSource<Type, Param> {
-  FutureOr<Type> update({Param param});
+  FutureOr<Type> update({required Param param});
+}
+
+/// Similar to [UpdaterDataSource], only returns Either form right away.
+abstract class SafeUpdaterDataSource<Type, Param> {
+  FutureOr<Either<Failure, Type>> update({required Param param});
 }
 
 /// Implement this abstract in your data source if it can perform data deletion.
 /// Note that 1 class can **implements** from multiple abstracts.
 ///
 /// ```dart
-/// class SessionDataSource implements DeleterDataSource<void, Null> {
+/// class SessionDataSource implements DeleterDataSource<Unit, Unit> {
 ///   final HttpClient client;
 ///
-///   const SessionDataSource({@required this.client});
+///   const SessionDataSource({required this.client});
 ///
+///   /// Use unit to represent void
 ///   @override
-///   FutureOr<void> delete({Null params}) async {
+///   FutureOr<Unit> delete({Unit params = unit}) async {
 ///     /// Perform sign out
 ///   }
 /// }
@@ -127,5 +151,10 @@ abstract class UpdaterDataSource<Type, Param> {
 ///
 /// P.S. Signing out means deleting session. Therefore, deleter data source.
 abstract class DeleterDataSource<Type, Param> {
-  FutureOr<Type> delete({Param param});
+  FutureOr<Type> delete({required Param param});
+}
+
+/// Similar to [DeleterDataSource], only returns Either form right away.
+abstract class SafeDeleterDataSource<Type, Param> {
+  FutureOr<Either<Failure, Type>> delete({required Param param});
 }